backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.1
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
SQLAlchemy==2.0.25
asyncpg==0.29.0
Pillow==10.2.0
alembic==1.18.4
pytest==8.3.5